requirements.txt
pydantic==2.8.2
dbt-core
dbt-postgres
dagster-dbt
dagster-docker
dagster_postgres
pandas
